feat: 로직분리 완료
