+++ src/hooks/auth/use-imge-preview.ts
@@ -1,0 +1,22 @@
+import { useState } from "react";
+
+export function useImagePreview() {
+  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
+
+  const handleImageSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    if (!file) return;
+
+    const reader = new FileReader();
+    reader.onloadend = () => {
+      setPreviewUrl(reader.result as string);
+    };
+    reader.readAsDataURL(file);
+  };
+
+  return {
+    previewUrl,
+    setPreviewUrl,
+    handleImageSelect,
+  };
+}
